requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
SQLAlchemy>=1.4.23
python-dotenv>=0.19.0
mysqlclient>=2.1.0
PyMySQL>=1.0.2
Pillow>=8.3.1
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.11.0
nltk>=3.6.3
python-magic>=0.4.24
aiofiles>=0.7.0
fastapi-clerk-auth==0.0.8
svix>=1.23.0